fix(UIController): 결과 출력 매개변수 double로 수정

+++ src/main/java/calculator/controller/UIController.java
@@ -12,11 +12,11 @@ public class UIController {
         return InputView.readValue();
     }
 
-    public static void printSumResult(int number) {
+    public static void printSumResult(double number) {
         OutputView.print(fitFormat(number));
     }
 
-    private static String fitFormat(int number) {
+    private static String fitFormat(double number) {
         return RESULT_FORMAT + number;
     }
 }
